removed `-pcf-resource-group` suffix.
this doesn't need to be there. if you use `pcf` as your `env_name`, you
get `pcf-pcf-resource-group`, and that name gets carried throughout the
rest of the resource groups. this will let you pick a name like `pcf`
and then everything is just referenced as `pcf-<resource>` throughout
the rest of the object creation process.

Signed-off-by: Mike Lloyd <mike@reboot3times.org>