support using cloud environments
This enables the terraform to use different cloud environments, such as Azure Germany.
The values are based of https://www.terraform.io/docs/providers/azurerm/#environment.
The variable `cloud_name` was used because `env` is already overloaded for OpsMan and PCF usage.
The default `cloud_name` is just using the US cloud environment, the original default of the terraform provider